Added Pathing.c & Pipe Function
